Removed IVariable interface
