--- Assets/Fungus/Scripts/Interfaces/IVariable.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 75ff90e0b811142d39cc9031dbf4b992
-timeCreated: 1473856441
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
